Mid-way through a clean-up

+++ Z80_Core/Bootstrapper.cs
@@ -8,7 +8,7 @@ namespace Z80.Core
     {
         public const int MAX_MEMORY_SIZE_IN_BYTES = 65536;
 
-        public static Processor BuildCPU(IMemoryMap map = null, ushort? topOfStackAddress = null, int speedInMHz = 4, bool enableFlagPrecalculation = true)
+        public static Processor BuildCPU(IMemoryMap map = null, ushort? topOfStackAddress = null, double speedInMHz = 4, bool enableFlagPrecalculation = true)
         {
             return new Processor(
                 map ?? new MemoryMap(MAX_MEMORY_SIZE_IN_BYTES, true),
